+++ src/Data/Export/EncoderInterface.php
@@ -9,7 +9,5 @@ use Remorhaz\JSON\Data\Value\ValueInterface;
 interface EncoderInterface extends ExporterInterface
 {
 
-    public function exportEvents(Iterator $eventIterator): string;
-
     public function exportValue(ValueInterface $value): string;
 }