+++ src/Data/Value/ArrayValueInterface.php
@@ -3,6 +3,6 @@ declare(strict_types=1);
 
 namespace Remorhaz\JSON\Data\Value;
 
-interface ArrayValueInterface extends ValueInterface
+interface ArrayValueInterface extends StructValueInterface
 {
 }