+++ src/Data/Value/DecodedJson/NodeValueFactory.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace Remorhaz\JSON\Data\Value\DecodedJson;
 
+use Remorhaz\JSON\Data\Iterator\ValueIteratorFactory;
+use Remorhaz\JSON\Data\Iterator\ValueIteratorFactoryInterface;
 use function is_array;
 use function is_scalar;
 use Remorhaz\JSON\Data\Path\Path;
@@ -13,9 +15,16 @@ use stdClass;
 final class NodeValueFactory implements NodeValueFactoryInterface
 {
 
+    private $valueIteratorFactory;
+
     public static function create(): NodeValueFactoryInterface
     {
-        return new self;
+        return new self(new ValueIteratorFactory);
+    }
+
+    public function __construct(ValueIteratorFactoryInterface $valueIteratorFactory)
+    {
+        $this->valueIteratorFactory = $valueIteratorFactory;
     }
 
     /**
@@ -35,11 +44,11 @@ final class NodeValueFactory implements NodeValueFactoryInterface
         }
 
         if (is_array($data)) {
-            return new NodeArrayValue($data, $path, $this);
+            return new NodeArrayValue($data, $path, $this, $this->valueIteratorFactory);
         }
 
         if ($data instanceof stdClass) {
-            return new NodeObjectValue($data, $path, $this);
+            return new NodeObjectValue($data, $path, $this, $this->valueIteratorFactory);
         }
 
         throw new Exception\InvalidNodeDataException($data, $path);