+++ src/Runtime/Comparator/EqualValueComparator.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 namespace Remorhaz\JSON\Path\Runtime\Comparator;
 
 use Collator;
-use Remorhaz\JSON\Data\Iterator\ValueIteratorFactoryInterface;
 use function is_string;
 use Remorhaz\JSON\Data\Value\ArrayValueInterface;
 use Remorhaz\JSON\Data\Value\ObjectValueInterface;
@@ -14,13 +13,10 @@ use Remorhaz\JSON\Data\Value\ValueInterface;
 final class EqualValueComparator implements ComparatorInterface
 {
 
-    private $valueIteratorFactory;
-
     private $collator;
 
-    public function __construct(ValueIteratorFactoryInterface $valueIteratorFactory, Collator $collator)
+    public function __construct(Collator $collator)
     {
-        $this->valueIteratorFactory = $valueIteratorFactory;
         $this->collator = $collator;
     }
 
@@ -55,8 +51,8 @@ final class EqualValueComparator implements ComparatorInterface
 
     private function isArrayEqual(ArrayValueInterface $leftValue, ArrayValueInterface $rightValue): bool
     {
-        $leftValueIterator = $this->valueIteratorFactory->createArrayIterator($leftValue->createEventIterator());
-        $rightValueIterator = $this->valueIteratorFactory->createArrayIterator($rightValue->createEventIterator());
+        $leftValueIterator = $leftValue->createChildIterator();
+        $rightValueIterator = $rightValue->createChildIterator();
 
         while ($leftValueIterator->valid()) {
             if (!$rightValueIterator->valid()) {
@@ -73,8 +69,8 @@ final class EqualValueComparator implements ComparatorInterface
 
     private function isObjectEqual(ObjectValueInterface $leftValue, ObjectValueInterface $rightValue): bool
     {
-        $leftValueIterator = $this->valueIteratorFactory->createObjectIterator($leftValue->createEventIterator());
-        $rightValueIterator = $this->valueIteratorFactory->createObjectIterator($rightValue->createEventIterator());
+        $leftValueIterator = $leftValue->createChildIterator();
+        $rightValueIterator = $rightValue->createChildIterator();
 
         $valuesByProperty = [];
         while ($leftValueIterator->valid()) {